qlxcx/django_back/requirements.txt
django-cors-headers
django
djangorestframework
mysqlclient
djangorestframework-simplejwt
Pillow 